调整 contracts-config 文件结构.添加部分注释
