+++ frontend/src/utils/index.js
@@ -3,10 +3,16 @@ const mainNetUrl = "https://confluxscan.io"
 
 const prefixMap = {
   1: testNetUrl,
-  1029: mainNetUrl
+  '1': testNetUrl,
+  1029: mainNetUrl,
+  '1029': mainNetUrl
 }
 
-// type: "transaction" or "address"
+/**
+ * @param content tx hash or account / contract address
+ * @param type "transaction" or "address"
+ * @param networkId 1 for testnet and 1029 for mainnet
+ * */ 
 function getScanUrl(content, type, networkId) {
   return [prefixMap[parseInt(networkId)], type, content].join('/')
 }
